Improve handling of conversion metrics (#173)
This builds on the work in #156 and #170 to apply similar improvements to Cumulative metrics.  

### General Context
Cumulative metrics need to have new metric inputs in place of their measure inputs, and if both inputs exist on the same model, the cumulative metric should also be merged into that model.

### What I'm doing here
Prior to 156, there were errors in general in making simple metrics to replace measure inputs when they had different values for fill_nulls_with or join_with_timespine.  This addresses that and more.  For a general idea of how this works, please review both 156 and 170 as this is based on that work entirely.

I've added a new test case to show that things work correctly for the top-level case.  I've also tried to thread comments throughout either in the code or in the PR, as appropriate, although some of them might get lost as commits are added.