Fix merging of dimensions (#184)
Dimension-merging gets complicated when we have expr values, and that's not represented yet in the autofix script.  We also didn't have a derived_semantics example in the tests to show that is working.  This fixes both things. :-)

=== Detailed Description

The conversations about this are spread between several places, so I'll try to condense it here.  First, in the original bug report, we see this behavior:

before:
```
dimensions:
      - name: order_total_dim
        type: categorical
        expr: order_total
```

currently turns into:
```
models: 
  - name: orders
    columns: 
	    ...
	          
      - name: order_total # no dims 
        description: The total amount of the order in USD including tax.

      - name: order_total_dim # not real
        dimension:
          type: categorical
```

when we'd expect to see:
```
models: 
  - name: orders
	  ...
    columns: 
	    ...
	          
      - name: order_total
        description: The total amount of the order in USD including tax.
        dimension: 
          type: categorical
          name: order_total_dim
```
@dave-connors-3 further clarified:
> If we can somehow map the expr to a 100 match on a column name, we should define it like you showed. idk how feasible that is, and if we can't, we should default it to the derived dimensions section to be safe


=== Code changes

I cleaned up the code for dimension merging a bit and addressed the behaviors when dimension has an expr. Then I added a test case to show that the derived semantics part was working, and I updated the test outcome for another dimension to show that it is merging based on the expr matching a column name now.